+++ spec/db/4.2/20171231235959_super_update.rb
@@ -1,0 +1,9 @@
+class SuperUpdate < ActiveRecord::Migration
+  def up
+    puts "Doing data migration"
+  end
+
+  def down
+    raise ActiveRecord::IrreversibleMigration
+  end
+end