+++ spec/db/data/20171231235959_super_update.rb
@@ -1,0 +1,9 @@
+class SuperUpdate < ActiveRecord::Migration[5.2]
+  def up
+    puts "Doing SuperUpdate"
+  end
+
+  def down
+    puts "Undoing SuperUpdate"
+  end
+end