+++ spec/db/migrate/5.0/20131111111111_late_migration.rb
@@ -1,0 +1,9 @@
+class LateMigration < ActiveRecord::Migration[5.0]
+  def up
+    puts "Doing schema LateMigration"
+  end
+
+  def down
+    puts "Undoing LateMigration"
+  end
+end